requirements.txt
appnope==0.1.0
astroid==1.5.3
bleach==2.1.1
CacheControl==0.12.4
cachetools==2.1.0
certifi==2017.11.5
chardet==3.0.4
click==6.7
Command==0.1.0
cycler==0.10.0
decorator==4.1.2
defusedxml==0.5.0
dj-database-url==0.4.2
docopt==0.6.2
entrypoints==0.2.3
ez-setup==0.9
firebase-admin==2.10.0
Flask==1.0.2
Flask-Cors==3.0.4
gcloud==0.17.0
google-api-core==1.1.2
google-api-python-client==1.6.7
google-auth==1.4.1
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.2.0
google-cloud-core==0.28.1
google-cloud-firestore==0.29.0
google-cloud-storage==1.10.0
google-resumable-media==0.3.1
googleapis-common-protos==1.5.3
grpcio==1.12.0
gtfs-realtime-bindings==0.0.5
gunicorn==19.8.1
html5lib==1.0b10
httplib2==0.11.3
idna==2.6
imageio==2.3.0
ipykernel==4.6.1
ipython==6.2.1
ipython-genutils==0.2.0
ipywidgets==7.0.3
isort==4.2.15
itsdangerous==0.24
jedi==0.11.0
Jinja2==2.10
jsonpickle==0.9.6
jsonschema==2.6.0
jws==0.1.3
lazy-object-proxy==1.3.1
lxml==4.1.1
Markdown==2.6.9
MarkupSafe==1.0
mccabe==0.6.1
mistune==0.7.4
moviepy==0.2.3.4
msgpack-python==0.5.6
mysql-connector-python==8.0.11
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.2.0
oauth2client==3.0.0
oauthlib==2.0.7
pandocfilters==1.4.2
parso==0.1.0
pdfminer.six==20170720
pexpect==4.2.1
pickleshare==0.7.4
Pillow==5.1.0
pipreqs==0.4.9
prompt-toolkit==1.0.15
protobuf==3.5.1
psycopg2==2.7.3.2
ptyprocess==0.5.2
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycryptodome==3.4.3
Pygments==2.2.0
PyJWT==1.6.1
pylint==1.7.4
PyMySQL==0.8.1
pyparsing==2.2.0
python-dateutil==2.6.1
python-jwt==2.0.1
python3-openid==3.1.0
pytz==2017.3
pyzmq==16.0.2
qtconsole==4.3.1
requests==2.18.4
requests-oauthlib==0.8.0
requests-toolbelt==0.7.0
rsa==3.4.2
simplegeneric==0.8.1
six==1.11.0
social-auth-app-django==2.1.0
social-auth-core==1.7.0
terminado==0.6
testpath==0.3.1
tornado==4.5.2
tqdm==4.23.0
traitlets==4.3.2
uritemplate==3.0.0
urllib3==1.22
virtualenv==15.2.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.0.6
wrapt==1.10.11
xlrd==1.1.0
yarg==0.1.9
